+++ Mozi.HttpEmbedded/Middleware/IMiddleware.cs
@@ -1,7 +1,10 @@
 ﻿namespace Mozi.HttpEmbedded.Middleaware
 {
+    /// <summary>
+    /// 中间件 接入多层次逻辑处理
+    /// </summary>
     public interface IMiddleware
     {
-
+        
     }
 }
