+++ Mozi.SSDP/Mozi.SSDP.csproj
@@ -36,6 +36,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="Control.cs" />
+    <Compile Include="Discovery.cs" />
     <Compile Include="Event.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RequestMethodUPnP.cs" />
